Updated products creation

+++ database/migrations/2019_09_24_140840_create_images_table.php
@@ -16,7 +16,7 @@ class CreateImagesTable extends Migration
         Schema::create('images', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name')->unique();
-            $table->decimal('size')->unsigned()->nullable();
+            $table->integer('size')->unsigned()->nullable();
 
             $table->timestamps();
         });
